Dataset
The columns include:

Gender: The type of cancer or category (e.g., "Any cancer", "Bladder", etc.).
Risk of developing (Male): The percentage risk and the equivalent "one in _ person" statistic. Risk of dying (Male): The percentage risk and the equivalent "one in person" statistic.
Risk of developing (Woman): The percentage risk and the equivalent "one in person" statistic. Risk of dying (Woman): The percentage risk and the equivalent "one in _ person" statistic.

Columns in the Dataset
Gender
Risk of developing (Male): Percentage
Risk of developing (Male): One in _ Person Risk of dying (Male): Percentage Risk of dying (Male): One in Person
Risk of developing (Woman): Percentage
Risk of developing (Woman): One in Person Risk of dying (Woman): Percentage Risk of dying (Woman): One in _ Person